refactor: Move config generation logic from Env to ServerMode
